requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-unfold==0.76.0
Faker==40.1.2
idna==3.11
mercadopago==2.3.0
pillow==12.1.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-unfold==0.76.0
Faker==40.1.2
idna==3.11
mercadopago==2.3.0
pillow==12.1.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
